RF | removing duplicate classes and moving code from them into one class
